+++ org.dawnsci.plotting.jreality/src/org/dawnsci/plotting/jreality/Multi2DTrace.java
@@ -49,7 +49,7 @@ public class Multi2DTrace extends Image3DTrace implements IMulti2DTrace{
 		
 		if (service==null) service = (IImageService)PlatformUI.getWorkbench().getService(IImageService.class);
 		if (rescaleHistogram) {
-			final float[] fa = service.getFastStatistics(imageServiceBean);
+			final double[] fa = service.getFastStatistics(imageServiceBean);
 			setMin(fa[0]);
 			setMax(fa[1]);
 		}
